BigQuery Analytics Pipeline for Financial Data/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.23.0
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
yfinance==0.2.26
pandas_ta==0.5.0b0
Pillow==9.5.0
